reference/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.3
fonttools==4.38.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
Naked==0.1.32
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pycryptodome==3.18.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
shellescape==3.8.1
six==1.16.0
urllib3==2.0.3
Werkzeug==2.2.3
zipp==3.13.0
